Implement lesson plan submit button functionality
Currently it checks lesson _resource_ form validity, but does not submit
updated file names and types. Currently, if Javascript is disabled on
the user's browser the button will still work. Since the user won't
be able to upload anything on the spot _anyway_, I think we're good on
that front.