Fix lesson resource uploading
This bug is easiest to understand if you imagine a lesson plan with one lesson
resource associated with it. Edit the lesson plan, remove the file and add a new
one. The frontend has no ID for this file, and will send an empty string for the
resourceIds hidden field, which Python cannot distinguish from no file. This
caused crashing at first, then the silent disregard of that new file. It took
some rework: JSON arrays are used for resourceIds, filetypes, and filenames and
a value of False is used for the ID of a NEW resource.